refactor: 共通のcalculateEndTimeStrをutilに移動
